move random functions into utils
